+++ src/test/kotlin/com/forgerock/uk/openbanking/tests/functional/payment/domestic/payments/consents/legacy/LegacyCreateDomesticPaymentsConsentsTest.kt
@@ -1,28 +1,27 @@
-package com.forgerock.uk.openbanking.tests.functional.payment.domestic.payments.consents
+package com.forgerock.uk.openbanking.tests.functional.payment.domestic.payments.consents.legacy
 
 import assertk.assertThat
 import assertk.assertions.contains
 import assertk.assertions.isEqualTo
 import assertk.assertions.isNotEmpty
 import assertk.assertions.isNotNull
-import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
 import com.forgerock.securebanking.framework.conditions.Status
 import com.forgerock.securebanking.framework.extensions.junit.CreateTppCallback
 import com.forgerock.securebanking.framework.extensions.junit.EnabledIfVersion
 import com.forgerock.securebanking.framework.http.fuel.defaultMapper
 import com.forgerock.securebanking.framework.signature.signPayloadSubmitPayment
 import com.forgerock.securebanking.framework.signature.signPayloadSubmitPaymentInvalidB64ClaimTrue
-import com.forgerock.uk.openbanking.support.discovery.payment3_1_1
-import com.forgerock.uk.openbanking.support.discovery.payment3_1_3
-import com.forgerock.uk.openbanking.support.discovery.payment3_1_4
-import com.forgerock.uk.openbanking.support.discovery.payment3_1_8
-import com.forgerock.uk.openbanking.support.payment.PaymentRS
+import com.forgerock.securebanking.openbanking.uk.common.api.meta.obie.OBVersion
 import com.forgerock.uk.openbanking.framework.constants.INVALID_FORMAT_DETACHED_JWS
 import com.forgerock.uk.openbanking.framework.constants.INVALID_SIGNING_KID
 import com.forgerock.uk.openbanking.framework.errors.INVALID_FORMAT_DETACHED_JWS_ERROR
 import com.forgerock.uk.openbanking.framework.errors.NO_DETACHED_JWS
 import com.forgerock.uk.openbanking.framework.errors.SIGNATURE_VALIDATION_FAILED
 import com.forgerock.uk.openbanking.framework.errors.UNAUTHORIZED
+import com.forgerock.uk.openbanking.support.discovery.payment3_1_1
+import com.forgerock.uk.openbanking.support.discovery.payment3_1_3
+import com.forgerock.uk.openbanking.support.discovery.payment3_1_4
+import com.forgerock.uk.openbanking.support.payment.PaymentRS
 import com.github.kittinunf.fuel.core.FuelError
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Assertions.assertThrows
@@ -30,171 +29,9 @@ import org.junit.jupiter.api.Test
 import uk.org.openbanking.datamodel.payment.OBWriteDomesticConsentResponse2
 import uk.org.openbanking.datamodel.payment.OBWriteDomesticConsentResponse3
 import uk.org.openbanking.datamodel.payment.OBWriteDomesticConsentResponse4
-import uk.org.openbanking.datamodel.payment.OBWriteDomesticConsentResponse5
 import uk.org.openbanking.testsupport.payment.OBWriteDomesticConsentTestDataFactory.*
 
-class CreateDomesticPaymentsConsentsTest(val tppResource: CreateTppCallback.TppResource) {
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateDomesticPaymentConsent"],
-        apis = ["domestic-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun createDomesticPaymentsConsents_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteDomesticConsent4()
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid
-            )
-
-        // When
-        val consent = PaymentRS().consent<OBWriteDomesticConsentResponse5>(
-            payment3_1_8.Links.links.CreateDomesticPaymentConsent,
-            consentRequest,
-            tppResource.tpp,
-            OBVersion.v3_1_8,
-            signedPayloadConsent
-        )
-
-        // Then
-        assertThat(consent).isNotNull()
-        assertThat(consent.data).isNotNull()
-        assertThat(consent.data.consentId).isNotEmpty()
-        Assertions.assertThat(consent.data.status.toString()).`is`(Status.consentCondition)
-        assertThat(consent.risk).isNotNull()
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateDomesticPaymentConsent"],
-        apis = ["domestic-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateDomesticPaymentsConsents_throwsSendInvalidFormatDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteDomesticConsent4()
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consent<OBWriteDomesticConsentResponse5>(
-                payment3_1_8.Links.links.CreateDomesticPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8,
-                INVALID_FORMAT_DETACHED_JWS
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
-        assertThat(exception.message.toString()).contains(INVALID_FORMAT_DETACHED_JWS_ERROR)
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateDomesticPaymentConsent"],
-        apis = ["domestic-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateDomesticPaymentsConsents_throwsNoDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteDomesticConsent4()
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consentNoDetachedJwt<OBWriteDomesticConsentResponse5>(
-                payment3_1_8.Links.links.CreateDomesticPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(400)
-        assertThat(exception.message.toString()).contains(NO_DETACHED_JWS)
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateDomesticPaymentConsent"],
-        apis = ["domestic-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateDomesticPaymentsConsents_throwsNotPermittedB64HeaderAddedInTheDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteDomesticConsent4()
-
-        val signedPayload =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                tppResource.tpp.signingKid,
-                true
-            )
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consent<OBWriteDomesticConsentResponse5>(
-                payment3_1_8.Links.links.CreateDomesticPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8,
-                signedPayload
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.responseMessage).isEqualTo(UNAUTHORIZED)
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(401)
-    }
-
-    @EnabledIfVersion(
-        type = "payments",
-        apiVersion = "v3.1.8",
-        operations = ["CreateDomesticPaymentConsent"],
-        apis = ["domestic-payment-consents"],
-        compatibleVersions = ["v.3.1.7", "v.3.1.6", "v.3.1.5"]
-    )
-    @Test
-    fun shouldCreateDomesticPaymentsConsents_throwsSendInvalidKidDetachedJws_v3_1_8() {
-        // Given
-        val consentRequest = aValidOBWriteDomesticConsent4()
-
-        val signedPayloadConsent =
-            signPayloadSubmitPayment(
-                defaultMapper.writeValueAsString(consentRequest),
-                tppResource.tpp.signingKey,
-                INVALID_SIGNING_KID
-            )
-
-        // When
-        val exception = assertThrows(AssertionError::class.java) {
-            PaymentRS().consent<OBWriteDomesticConsentResponse5>(
-                payment3_1_8.Links.links.CreateDomesticPaymentConsent,
-                consentRequest,
-                tppResource.tpp,
-                OBVersion.v3_1_8,
-                signedPayloadConsent
-            )
-        }
-
-        // Then
-        assertThat((exception.cause as FuelError).response.statusCode).isEqualTo(401)
-        assertThat((exception.cause as FuelError).response.responseMessage).isEqualTo(UNAUTHORIZED)
-    }
+class LegacyCreateDomesticPaymentsConsentsTest(val tppResource: CreateTppCallback.TppResource) {
 
     @EnabledIfVersion(
         type = "payments",
